+++ src/api/UserApi.tsx
@@ -15,11 +15,6 @@ const UserApi = {
     axios.post("login", { email, password }),
   logout: () => axios.post("logout"),
   join: (obj: JoinObject) => axios.post("/join", obj),
-  enroll: (name: string, age: number, gender: string, phone: string) =>
-    axios.post("customers", { name, age, gender, phone }),
-  assign: (trainerId: string, customerId: string) =>
-    axios.post("/customers/:trainerId/:customerId", { trainerId, customerId }),
-  getUserAll: () => axios.get("/customers"),
 };
 
 export default UserApi;
