+++ src/api/index.tsx
@@ -1,2 +1,3 @@
 export { default as UserApi } from "./UserApi";
 export { default as TrainerApi } from "./TrainerApi";
+export { default as CustomerApi } from "./CustomerApi";
